CheckBox is fully declarative
CheckBox checked state is only set by prop.
